% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.na.R
\name{count_missing_values}
\alias{count_missing_values}
\title{An R package to count missing values in a dataset
The function takes the dataframe and produces a sum of missing values}
\usage{
count_missing_values(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
A vector
}
\description{
An R package to count missing values in a dataset
The function takes the dataframe and produces a sum of missing values
}
\examples{
df <- data.frame(x = c(1, 2, NA), y = c(NA, 5, 6))
 count_missing_values(df)
}
